// txcnotes/source/TXC-MicroFrontends/Order/src/app/shared/enums/template.enum.ts
export enum TemplateTypeEnum {
  Email = 1,
  SMS = 2,
  Other = 3
}

export enum TemplateSubTypeEnum {
  Voucher = 1,
  SMS = 2,
  Other = 3
}

export enum TagTypeEnum {
  Text = 1,
  Image = 2,
  RichText = 3,
  RadioGroup = 4,
  HTMLText = 5,
}

export enum TagCategoryEnum {
  System = 1,
  UserInput = 2,
  UserInputWithDefault = 3,
  SelectedBySystem = 4,
}